+++ stories/components/base-components/scroll-area/tests/examples/scroll-props.tsx
@@ -12,10 +12,6 @@ export type ScrollAreaExampleProps = ScrollAreaProps & {
   shadowSize?: number;
   itemsCount: number;
   shadowTheme?: 'dark' | 'light';
-  focusRingTopOffset?: string;
-  focusRingRightOffset?: string;
-  focusRingBottomOffset?: string;
-  focusRingLeftOffset?: string;
 };
 
 export const defaultProps: ScrollAreaExampleProps = {
@@ -51,7 +47,7 @@ const ScrollAreaDemo = ({
   ...rest
 }: ScrollAreaExampleProps) => {
   return (
-    <Flex p={4} style={{ backgroundColor: '#F5F6FA', minHeight: 360 }}>
+    <Flex p={4} bg='#F5F6FA' hMin={360}>
       <ScrollArea
         w={300}
         h={300}