+++ stories/components/progress-bar/docs/examples/customizing_the_bar.tsx
@@ -6,15 +6,15 @@ type ProgressBarExampleProps = NSProgressBar.Props;
 const Demo = (props: ProgressBarExampleProps) => {
   return (
     <ProgressBar tabIndex={0} value={props.value} size={props.size} theme={props.theme} aria-label='Custom theme example' duration={props.duration}>
-      <ProgressBar.Value theme='violet-500' />
+      <ProgressBar.Value theme='brand-secondary' />
     </ProgressBar>
   );
 };
 
 export const defaultProps: ProgressBarExampleProps = {
   size: 'm',
-  value: 80,
-  theme: 'violet-100',
+  value: 60,
+  theme: 'progress-bar-bg',
   duration: undefined,
 };